requirements.txt
pandas>=2.1.0
python-dotenv==1.0.0
google-genai>=1.38.0
numpy>=1.24.0
openai>=1.56.0